% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessels.R
\name{get_latest_location}
\alias{get_latest_location}
\title{Get Latest Vessel Transmission}
\usage{
get_latest_location(api_key, mmsi)
}
\arguments{
\item{api_key, }{API access token}

\item{mmsi, }{mmsi tied to the vessel to get history trail for}
}
\description{
returns latest vessel transmission from history trail
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
