% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessels.R
\name{get_location_vessels}
\alias{get_location_vessels}
\title{Get Location Vessels}
\usage{
get_location_vessels(api_key, latitude, longitude, age = NULL, radious = NULL)
}
\arguments{
\item{api_key, }{API access token}

\item{latitude, }{latitude of the center point for the region of interest}

\item{longitude, }{longitude of the center point for the region of interest}

\item{age, }{how old the vessels that are returned can be in hours(must be between 1 and 720 hours)}

\item{radious, }{number of miles to search for vessels (must be 100 miles or under )}
}
\description{
returns Vessels within a 100 mile radius of a specified latitude and longitude
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
