% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessels.R
\name{get_vessel_attributes}
\alias{get_vessel_attributes}
\title{Get Vessel Attributes}
\usage{
get_vessel_attributes(api_key, mmsi)
}
\arguments{
\item{api_key, }{API access token}

\item{mmsi, }{mmsi tied to the vessel to get history trail for}
}
\description{
returns ship attributes (e.g., mmsi, imoNumber, name, callSign, cargo, vesselType, COG, heading, navStatus, SOG, latitude, longitude, timeOfFix, lenght, beam) based on the vessels last location
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
